Commit 39.5
Bugs corrigidos. levar hit do inimigo não mata o player instantaneamente agora...

+++ Assets/Scripts/Scripts Sample Scene/Power Ups/Vidaextra.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class Vidaextra : MonoBehaviour
 {
-   
-    public float _vidaParaDar;
+
+    public int _vidaParaDar;
 
     private void OnTriggerEnter(Collider other)
     {
